Refactor keccak + improve testing
